# Some small optimisations