Reduces one allocation in `provideFor` and a few other things in paths of varying temperatures (notably getting rid of `fmt.Sprintf` which isn't the most efficient way to put strings together, although it is fairly concise).
Before:
```
INFO: Run 1 of 5
INFO: Complete in 9.62s, using 6172140 KB
INFO: Run 2 of 5
INFO: Complete in 9.18s, using 5655136 KB
INFO: Run 3 of 5
INFO: Complete in 9.62s, using 6437840 KB
INFO: Run 4 of 5
INFO: Complete in 9.68s, using 6682404 KB
INFO: Run 5 of 5
INFO: Complete in 9.61s, using 6525640 KB
INFO: Complete, median time: 9.62s, median mem: 6437840.00 KB
```
After:
```
INFO: Run 1 of 5
INFO: Complete in 9.42s, using 6408832 KB
INFO: Run 2 of 5
INFO: Complete in 9.30s, using 6354172 KB
INFO: Run 3 of 5
INFO: Complete in 9.18s, using 6287200 KB
INFO: Run 4 of 5
INFO: Complete in 9.28s, using 5919368 KB
INFO: Run 5 of 5
INFO: Complete in 9.40s, using 6682712 KB
INFO: Complete, median time: 9.30s, median mem: 6354172.00 KB
```