@peterebden 2024-10-03
> What benchmark are you running in your before/after?

The performance one we run after merge (`//tools/performance:parse_perf_test`)